network: Improve context creation robustness
It might happen that the uri passed to libiio is an avahi typical
hostname. In that case, is not guaranteed that `getaddrinfo()` will
succeed in resolving it. For that to happen there are dependencies
on the host configuration like the presence of the nss-mdns package.
This patch adds a "a last resort" attempt in resolving the hostname in
case `getaddrinfo()` fails, by directly using libavahi.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>